requirements.txt
django==3.2.12
djangorestframework
django-filter
dj-rest-auth